pyproject.toml
[project]
name = "ppt-agent"
version = "0.1.0"
description = "AI-powered PowerPoint deck generator using LangGraph and LangChain"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "langchain>=1.1.0",
    "langgraph>=1.0.6",
    "langchain-openai>=0.3.0",
    "langsmith>=0.6.3",
    "python-pptx>=1.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=24.0.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
# UV-specific configuration
dev-dependencies = [
    "pytest>=7.0.0",
    "black>=24.0.0",
    "ruff>=0.1.0",
]

[tool.black]
line-length = 100
target-version = ['py310']

[tool.ruff]
line-length = 100
target-version = "py310"
